+++ release.sh
@@ -39,7 +39,7 @@ else
 fi
 
 # tag all images and push if needed
-for i in clowder toolserver mongo-init; do
+for i in clowder toolserver mongo-init monitor; do
     for v in ${VERSION}; do
         if [ "$v" != "latest" -o "$SERVER" != "" ]; then
             ${DEBUG} docker tag clowder/${i}:latest ${SERVER}clowder/${i}:${v}